+++ client/src/components/UI/organisms/NotificationContainer/index.tsx
@@ -1,9 +1,8 @@
-import React, { FC, useEffect, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
+import React, { FC } from 'react';
 import styled from 'styled-components';
 
-import CloseIcon from '@icons/CloseIcon';
 import { Notification } from '@organisms/header';
+import NotificationItem from '@molecules/NotificationItem';
 
 const StyledNotificationContainer = styled.div`
     &::before {
@@ -40,91 +39,11 @@ const StyledHeader = styled.h3`
     padding: 1rem;
 `;
 
-const StyledNotificationItem = styled.li`
-    list-style: none;
-    height: 3rem;
-    padding: 1rem;
-    font-size: 75%;
-    border-top: 1px solid rgba(0, 0, 0, 0.1);
-
-    i {
-        color: #b5c4d2;
-        font-size: 140%;
-        position: absolute;
-        &.right {
-            right: 1rem;
-            &:hover {
-                opacity: 0.8;
-                cursor: pointer;
-            }
-        }
-    }
-
-    em {
-        width: 3em;
-        margin-right: 1rem;
-        font-weight: 700;
-        font-size: 115%;
-        color: #b5c4d2;
-        vertical-align: bottom;
-        display: inline-block;
-    }
-
-    .notification {
-        box-sizing: border-box;
-        &.new {
-            background: #f3f9fd;
-        }
-    }
-`;
-
-interface CreatedRoomNotification {
-    type: string;
-    roomId: string;
-}
-
-interface OpenedTicketNotification {
-    type: string;
-    ticketId: string;
-}
-
-const NotificationItem = ({ notification, id }: any) => {
-    const navigate = useNavigate();
-
-    // TODO : Link 태그?
-    const navigateByAlarmType = (type: string, id: string) => {
-        navigate(`/${type}/${id}`);
-    };
-
-    return (
-        <StyledNotificationItem
-            className="notification new"
-            onClick={() => navigateByAlarmType(notification.type, notification.id)}
-        >
-            <em>오늘</em> {notification.message}
-            <i className="right">
-                <CloseIcon />
-            </i>
-        </StyledNotificationItem>
-    );
-};
-
 interface Props {
     notifications: Notification[];
 }
 
-const testNotification = {
-    type: 'ticket',
-    id: '1',
-};
-
 const NotificationContainer: FC<Props> = ({ notifications }) => {
-    const navigate = useNavigate();
-
-    const navigateByAlarmType = (type: string, id: string) => {
-        navigate(`/${type}/${id}`);
-    };
-
     return (
         <StyledNotificationContainer>
             <StyledHeader>Notifications</StyledHeader>
